LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/getactionurl-expected.txt

CONSOLE MESSAGE: line 1380: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.



FAIL Navigating to URL with a data scheme undefined is not an object (evaluating 'testframe.contentWindow.document.textContent')
